+++ src/manipulation/transformSpanToRange.cpp
@@ -35,10 +35,10 @@ Range *transformSpanToRange(Range *r, hif::semantics::ILanguageSemantics *refLan
         IntValue *i = static_cast<IntValue *>(spanDimension);
         if (isSigned) {
             range->setRightBound(
-                new IntValue(static_cast<long long>(pow(2, static_cast<double>(i->getValue()) - 1) - 1)));
-            range->setLeftBound(new IntValue(static_cast<long long>(-pow(2, static_cast<double>(i->getValue()) - 1))));
+                new IntValue(static_cast<std::int64_t>(pow(2, static_cast<double>(i->getValue()) - 1) - 1)));
+            range->setLeftBound(new IntValue(static_cast<std::int64_t>(-pow(2, static_cast<double>(i->getValue()) - 1))));
         } else {
-            range->setRightBound(new IntValue(static_cast<long long>(pow(2, static_cast<double>(i->getValue())) - 1)));
+            range->setRightBound(new IntValue(static_cast<std::int64_t>(pow(2, static_cast<double>(i->getValue())) - 1)));
             range->setLeftBound(new IntValue(0));
         }
     } else {
